requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
asyncio==3.4.3
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
discord.py==0.16.0
multidict==4.5.2
pycparser==2.19
PyNaCl==1.0.1
six==1.12.0
websockets==3.4
wincertstore==0.2
